Move eh_personality() onto CrateContext
